+++ example/signature.dart
@@ -1,7 +1,7 @@
 import 'package:pinenacl/ed25519.dart';
 
 void main() {
-  const hex = HexCoder.instance;
+  const hex = Base16Encoder.instance;
   print('\n### Digital Signatures - Signing Example ###\n');
 
   /// Signer’s perspective (SigningKey)