+++ lib/api/encoding.dart
@@ -1,5 +1,7 @@
 part of pinenacl.api;
 
+/// The Encoder interface for classes that are capable for encoding data,
+/// therefore they need decoding function too.
 abstract class Encoder {
   String encode(ByteList data);
   Uint8List decode(String data);