+++ lib/encoding.dart
@@ -8,4 +8,4 @@ import 'package:pinenacl/api.dart';
 
 part 'src/encoding/base32_encoder.dart';
 part 'src/encoding/bech32_encoder.dart';
-part 'src/encoding/hex_encoder.dart';
+part 'src/encoding/base16_encoder.dart';